src/requirements.txt
alembic==1.12.0
annotated-types==0.6.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
greenlet==3.0.0
idna==3.4
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
pipdeptree==2.13.0
protobuf==4.21.12
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
soupsieve==2.5
SQLAlchemy==2.0.22
SQLAlchemy-Utils==0.41.1
tenacity==8.2.3
typing_extensions==4.8.0
urllib3==2.0.7
